% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptGraph.R
\name{ScriptGraph}
\alias{ScriptGraph}
\title{This script calculates the IDF-Graph}
\usage{
ScriptGraph(Directory)
}
\arguments{
\item{Directory}{Input Location where the IDF equations}
}
\description{
This script calculates the IDF-Graph
}
