% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptGraphic.R
\name{ScriptGraphic}
\alias{ScriptGraphic}
\title{This script generates the graph of the intensities}
\usage{
ScriptGraphic(Directory)
}
\arguments{
\item{Directory}{Parameters IDF Equation}
}
\description{
This script generates the graph of the intensities
}
