% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptIDFCLIMBra.R
\name{ScriptIDFCLIMBra}
\alias{ScriptIDFCLIMBra}
\title{This script calculates the IDF equations - CLIMBra}
\usage{
ScriptIDFCLIMBra(CLIMBra, Station, LatSR, LonSR, Directory)
}
\arguments{
\item{CLIMBra}{Type GCMs}

\item{Station}{Input Name station}

\item{LatSR}{LonSR Input Latitude}

\item{LonSR}{Input Longitude}

\item{Directory}{Location where the IDF equations will be saved}
}
\description{
This script calculates the IDF equations - CLIMBra
}
