% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptIDFSR.R
\name{ScriptIDFSR}
\alias{ScriptIDFSR}
\title{This script calculates the IDF equations - remote sensing}
\usage{
ScriptIDFSR(Station, LatSR, LonSR, Directory)
}
\arguments{
\item{Station}{Input Name station}

\item{LatSR}{LonSR Input Latitude}

\item{LonSR}{Input Longitude}

\item{Directory}{Location where the IDF equations will be saved}
}
\description{
This script calculates the IDF equations - remote sensing
}
