% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptIDF.R
\name{ScriptIDF}
\alias{ScriptIDF}
\title{This script calculates the IDF equations for any place in Brazil}
\usage{
ScriptIDF(StatesANA, Directory)
}
\arguments{
\item{StatesANA}{Input data server ANA or local}

\item{Directory}{Location where the IDF equations will be saved}
}
\description{
This script calculates the IDF equations for any place in Brazil
}
