% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptIDF.R
\name{ScriptIDF}
\alias{ScriptIDF}
\title{This script calculates the IDF Equations for Brazil}
\usage{
ScriptIDF(StatesANA, Directory, Method, Isozone)
}
\arguments{
\item{StatesANA}{Input Data server ANA ("ACRE") or Local ("")}

\item{Directory}{Input Location where the IDF equations will be saved}

\item{Method}{Input Disaggregation method}

\item{Isozone}{Input Isozone}
}
\description{
This script calculates the IDF Equations for Brazil
}
