% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptIDFCLIMBra.R
\name{ScriptIDFCLIMBra}
\alias{ScriptIDFCLIMBra}
\title{This script calculates the IDF Equations - CLIMBra}
\usage{
ScriptIDFCLIMBra(CLIMBra, Station, LatSR, LonSR, Directory, Method, Isozone)
}
\arguments{
\item{CLIMBra}{Input Type GCMs (ssp245 or ssp585)}

\item{Station}{Input Name station}

\item{LatSR}{Input Latitude}

\item{LonSR}{Input Longitude}

\item{Directory}{Input Location where the IDF equations will be saved#'}

\item{Method}{Input Disaggregation method}

\item{Isozone}{Input Isozone}
}
\description{
This script calculates the IDF Equations - CLIMBra
}
