% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptIDFSR.R
\name{ScriptIDFSR}
\alias{ScriptIDFSR}
\title{This script calculates the IDF Equations - Remote-sensing}
\usage{
ScriptIDFSR(Station, LatSR, LonSR, Directory, Method, Isozone)
}
\arguments{
\item{Station}{Input Name station}

\item{LatSR}{Input Latitude}

\item{LonSR}{Input Longitude}

\item{Directory}{Input Location where the IDF equations will be saved}

\item{Method}{Input Disaggregation method}

\item{Isozone}{Input Isozone}
}
\description{
This script calculates the IDF Equations - Remote-sensing
}
